requirements.txt
google-generativeai
python-dotenv
Pillow
requests